trying

+++ occtet-nats/occtet-nats-issue-catcher-service/src/main/java/eu/occtet/boc/issueCatcher/service/IssueCatcherService.java
@@ -15,7 +15,7 @@ public class IssueCatcherService {
         return fetchIssues(workData.getRunId());
     }
 
-    public boolean fetchIssues(int runId){
+    public boolean fetchIssues(long runId){
             String jsonResponse="";
             try {
                 //use ORT API to fetch issues of a ort run
